Refactor Builder pattern, adopt SafeSpan, and fix TcpServer shutdown (#210)
* Refactor Builder pattern, adopt SafeSpan, and fix TcpServer shutdown

- Implemented CRTP for Builder pattern to remove manual casting.
- Adopted `std::string_view` in `common.hpp`, `logger.hpp`, and `error_handler.hpp`.
- Enforced `SafeSpan` usage in all transport interfaces and implementations (`TcpServer`, `TcpClient`, `Udp`, `Serial`).
- Refactored `TcpServer::stop()` to ensure proper shutdown order (acceptor first) and robust cleanup via `net::dispatch`.
- Fixed `Serial` wrapper to remove code duplication in `start()` and restore case-insensitivity.
- Updated UDP tests to handle timing and backpressure limits correctly.


* Refactor architecture, performance, and safety; fix CI failures

- Implement Builder pattern CRTP for fluent API.
- Adopt std::string_view for performance.
- Enforce SafeSpan usage in transport layer.
- Improve TcpServer shutdown stability.
- Fix CI compilation errors (python bindings, unit tests, performance tests).
- Fix warnings treated as errors (-Wsign-conversion, -Wunused-parameter, etc).
- Fix duplicate initialization in TcpServer constructor.
- Fix Python bindings on_bytes signature.


* Fix CI failures: formatting, SafeSpan migration in tests, and race conditions

- Updated `transport_tcp_fuzz_test.cc` to use `ConstByteSpan` in `on_bytes` callback, resolving compilation error.
- Ran `clang-format` on all files flagged by CI check.
- Fixed race conditions in `TransportTcpFuzzTest`, `TransportTcpServerSessionTest`, and `TransportTcpTimeoutTest` by ensuring `io_context` has work (using `make_work_guard`) and waiting for handlers to be registered before emitting mock data.
- Added `has_handler()` to `FakeTcpSocket` to support robust testing synchronization.
- Added `override` to destructors in `ConfigManager` and `TcpServer` to fix compiler warnings.


* Fix CI failures: formatting, SafeSpan migration, and test stability

- Updated `transport_tcp_fuzz_test.cc` to use `ConstByteSpan` in `on_bytes` callback, resolving compilation error.
- Ran `clang-format` on all files flagged by CI check.
- Fixed race conditions in `TransportTcpFuzzTest`, `TransportTcpServerSessionTest`, and `TransportTcpTimeoutTest` by ensuring `io_context` has work (using `make_work_guard`) and waiting for handlers to be registered before emitting mock data.
- Added `has_handler()` to `FakeTcpSocket` to support robust testing synchronization.
- Added `override` to destructors in `ConfigManager` and `TcpServer` to fix compiler warnings.
- Resolved merge conflicts by combining changes from both branches.


* Fix CI duplicate target errors and formatting

- Remove explicit add_executable calls for targets already covered by foreach loop in test/unit/CMakeLists.txt to fix Policy CMP0002 errors.
- Fix clang-format violations in transport unit tests and benchmark tests.
- Update FakeTcpSocket to implement missing virtual methods and improve callback thread safety.
- Harden transport unit tests against race conditions by using work_guard and ensuring handlers are registered.


* Fix MSVC CI build errors by removing semicolon-separated lists in CMake LABELS

MSVC argument parsing on CI was failing when semicolon-separated lists were passed as LABELS properties in `gtest_discover_tests`. This commit replaces these lists with underscore-separated single strings (e.g., `unit;common;fast` -> `unit_common_fast`) in both `test/unit/CMakeLists.txt` and `test/integration/CMakeLists.txt` to ensure compatibility across all platforms.


---------

+++ test/performance/benchmark/test_memory_tracker_perf.cc
@@ -38,18 +38,18 @@ class MemoryTrackerPerfTest : public ::testing::Test {
 };
 
 TEST_F(MemoryTrackerPerfTest, TrackAllocationOverhead) {
-  const int iterations = 1000000;
+  const size_t iterations = 1000000;
   std::vector<void*> ptrs;
   ptrs.reserve(iterations);
 
   // Pre-allocate pointers to avoid measuring vector allocation time
-  for (int i = 0; i < iterations; ++i) {
+  for (size_t i = 0; i < iterations; ++i) {
     ptrs.push_back(reinterpret_cast<void*>(static_cast<uintptr_t>(i + 1)));
   }
 
   auto start = std::chrono::high_resolution_clock::now();
 
-  for (int i = 0; i < iterations; ++i) {
+  for (size_t i = 0; i < iterations; ++i) {
     MemoryTracker::instance().track_allocation(ptrs[i], 64, __FILE__, __LINE__, __FUNCTION__);
   }
 
@@ -60,17 +60,17 @@ TEST_F(MemoryTrackerPerfTest, TrackAllocationOverhead) {
   std::cout << "Time per op: " << (double)duration * 1000000 / iterations << " ns" << std::endl;
 
   // Cleanup to avoid memory usage growth during test
-  for (int i = 0; i < iterations; ++i) {
+  for (size_t i = 0; i < iterations; ++i) {
     MemoryTracker::instance().track_deallocation(ptrs[i]);
   }
 }
 
 TEST_F(MemoryTrackerPerfTest, TrackDeallocationOverhead) {
-  const int iterations = 1000000;
+  const size_t iterations = 1000000;
   std::vector<void*> ptrs;
   ptrs.reserve(iterations);
 
-  for (int i = 0; i < iterations; ++i) {
+  for (size_t i = 0; i < iterations; ++i) {
     void* ptr = reinterpret_cast<void*>(static_cast<uintptr_t>(i + 1));
     ptrs.push_back(ptr);
     MemoryTracker::instance().track_allocation(ptr, 64, __FILE__, __LINE__, __FUNCTION__);
@@ -78,7 +78,7 @@ TEST_F(MemoryTrackerPerfTest, TrackDeallocationOverhead) {
 
   auto start = std::chrono::high_resolution_clock::now();
 
-  for (int i = 0; i < iterations; ++i) {
+  for (size_t i = 0; i < iterations; ++i) {
     MemoryTracker::instance().track_deallocation(ptrs[i]);
   }
